% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airac-package.R
\docType{package}
\name{airac-package}
\alias{airac-package}
\alias{_PACKAGE}
\title{airac: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
