% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airac.R
\name{airac}
\alias{airac}
\title{Return the ICAO AIRAC id for a date}
\usage{
airac(date)
}
\arguments{
\item{date}{a date (or a representation of it understood by \code{\link[lubridate:ymd]{lubridate::ymd()}})}
}
\value{
the ICAO AIRAC id
}
\description{
Return the ICAO AIRAC id for a date
}
\examples{
airac('2019-01-30')

}
\seealso{
Other airac: 
\code{\link{airac_epoch}()},
\code{\link{airac_interval}()},
\code{\link{airac_year_epoch}()},
\code{\link{cfmu_airac_epoch_number}()},
\code{\link{cfmu_airac_epoch}()},
\code{\link{cfmu_airac_interval}()},
\code{\link{cfmu_airac}()}
}
\concept{airac}
