% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airac.R
\name{airac_interval}
\alias{airac_interval}
\title{The interval of dates for an (ICAO) AIRAC}
\usage{
airac_interval(airac)
}
\arguments{
\item{airac}{the ICAO AIRAC, i.e. "1603" for the 3rd AIRAC of 2016}
}
\value{
an interval of (28) days covered by the AIRAC \code{airac}
}
\description{
The interval of dates for an (ICAO) AIRAC
}
\examples{
airac_interval("2403")

}
\seealso{
Other airac: 
\code{\link{airac_epoch}()},
\code{\link{airac_year_epoch}()},
\code{\link{airac}()},
\code{\link{cfmu_airac_epoch_number}()},
\code{\link{cfmu_airac_epoch}()},
\code{\link{cfmu_airac_interval}()},
\code{\link{cfmu_airac}()}
}
\concept{airac}
