% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airac-package.R
\docType{package}
\name{airac-package}
\alias{airac-package}
\alias{_PACKAGE}
\title{airac: Helper functions to get ICAO AIRAC ID and intervals as used in Aeronautical Information Services and Publications.}
\description{
Derive AIRAC ID and relevant intervals both as used internationally and by the data services of the European Network Manager, i.e. the CFMU AIRAC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eurocontrol/airac}
  \item \url{https://eurocontrol.github.io/airac/}
  \item Report bugs at \url{https://github.com/eurocontrol/airac/issues}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

Other contributors:
\itemize{
  \item EUROCONTROL [copyright holder, funder]
}

}
\keyword{internal}
